Change timer_get_seconds to double for precision

+++ tests/timer_test.c
@@ -3,22 +3,22 @@
 
 int main(int argc, char **argv)
 {
-  int timer_limit;
+  double timer_limit;
   if(argc <= 1)
   {
     printf("No time limit specified, defaulting to 3 seconds\n");
-    timer_limit = 3;
+    timer_limit = 3.00;
   } else
   {
-    timer_limit = (int)(*argv[1] - '0');
+    timer_limit = (double)(*argv[1] - '0');
   }
 
   timer_reset();
-  int seconds = 0;
+  double seconds = 0;
   do
   {
     seconds = timer_get_seconds();
-    printf("\rElapsed seconds: %d / %d", seconds, timer_limit);
+    printf("\rElapsed seconds: %f / %f", seconds, timer_limit);
   } while (seconds < timer_limit);
   return 0;
 }
